added warning message for delete option

+++ src/app/(dashboard)/dashboard/supplier/AddSurcharge/page.tsx
@@ -339,6 +339,8 @@ const Surcharge = () => {
   };
 
   const handleDelete = async (id: string) => {
+    const confirmDelete = window.confirm("Are you sure you want to delete?");
+  if (!confirmDelete) return;
     try {
       await fetchWithAuth(`${API_BASE_URL}/supplier/DeleteSurgeCharges/${id}`, {
         method: "DELETE",
